requirements.txt
anyio==4.0.0
certifi==2023.7.22
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
isort==5.12.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
six==1.16.0
sniffio==1.3.0
